refactor: split channel-guardian-store.ts into focused modules
